fix: Invalid writeVarInt for 32 bit numbers

+++ packages/contracts/contracts/illuminex/Buffer.sol
@@ -94,7 +94,7 @@ library Buffer {
             write(buffer, bytes.concat(bytes2(Endian.reverse16(uint16(value)))));
         } else if (value >= 65536 && value <= 0xFFFFFFFF) {
             write(buffer, bytes.concat(bytes1(uint8(0xFE))));
-            write(buffer, bytes.concat(bytes4(Endian.reverse32(uint16(value)))));
+            write(buffer, bytes.concat(bytes4(Endian.reverse32(uint32(value)))));
         } else if (value >= 4_294_967_296 && value <= 0xFFFFFFFFFFFFFFFF) {
             write(buffer, bytes.concat(bytes1(uint8(0xFF))));
             write(buffer, bytes.concat(bytes8(Endian.reverse64(uint64(value)))));
